Move static methods implementations out from DBUtils and deprecated them
